--- protos/face_label_map.pbtxt
@@ -1,9 +1,0 @@
-item {
-  id: 2
-  name: 'background'
-}
-
-item {
-  id: 1
-  name: 'face'
-}
