+++ src/run_camera
@@ -1,0 +1,12 @@
+#!/usr/bin/env sh
+
+cd $SNAP
+
+MODEL_DIR="$SNAP/model"
+export MODEL_DIR
+PROTOS_DIR="$SNAP/protos"
+export PROTOS_DIR
+
+touch $SNAP_DATA/i3.config
+
+python3 $SNAP/inference_usbCam_face.py 0
