+++ src/run_video
@@ -1,0 +1,11 @@
+#!/usr/bin/env sh
+
+cd $SNAP
+
+MODEL_DIR="$SNAP/model"
+export MODEL_DIR
+PROTOS_DIR="$SNAP/protos"
+export PROTOS_DIR
+
+python3 $SNAP/inference_video_face.py "$@"
+
