+++ src/setup.py
@@ -1,0 +1,6 @@
+from setuptools import setup, find_packages
+setup(
+    name="HelloWorld",
+    version="0.1",
+    packages=find_packages(),
+)
